userList

+++ FInalProject_Health/src/main/java/com/health/controller/JournalController.java
@@ -3,10 +3,13 @@ package com.health.controller;
 import java.io.PrintWriter;
 import java.sql.Timestamp;
 import java.util.ArrayList;
+import java.util.Enumeration;
 import java.util.List;
 import java.util.Map;
 
+import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -20,6 +23,7 @@ import org.springframework.web.bind.annotation.RequestParam;
 
 import com.health.domain.Exercise;
 import com.health.domain.Journal;
+import com.health.domain.Member;
 import com.health.service.JournalService;
 
 @Controller
@@ -27,54 +31,49 @@ public class JournalController {
 
 	@Autowired
 	private JournalService service;
+	 
+	@PostMapping("/eWriteProcess") 
+	public String insertExercise(
+			HttpServletRequest request, HttpSession session, 
+			@RequestParam(value="exerciseCate", required=false)String[] exerciseCates, 
+			@RequestParam(value="exerciseName", required=false)String[] exerciseNames,
+	        @RequestParam(value="journalNo", required=false) int journalNo, 
+	        @RequestParam(value="exerciseWeight", required=false) int[] exerciseWeights, 
+	        @RequestParam(value="exerciseSet", required=false) int[] exerciseSets, 
+	        @RequestParam(value="exerciseRep", required=false) int[] exerciseReps) { 
+		
+		session = request.getSession();
+		Member member = (Member) session.getAttribute("member");
+		if(member == null){
+			return "redirect:loginForm";
+		}
+	    for(int i=0; i <exerciseCates.length; i++) {
+		Exercise exercise = new Exercise();
+	    exercise.setExerciseCate(exerciseCates[i]);
+	    exercise.setExerciseName(exerciseNames[i]);
+	    exercise.setExerciseWeight(exerciseWeights[i]);
+	    exercise.setExerciseSet(exerciseSets[i]); 
+	    exercise.setExerciseRep(exerciseReps[i]);
+	    exercise.setFK_exercise_journal(journalNo);
+	    service.insertExercise(exercise);
+	    }
+	   
+	    return "redirect:/journalDetail?journalNo=" + journalNo; 
+	}
 
 	
-	
-	/*@PostMapping("/eWriteProcess") 
-	public String insertExercise(
-			@RequestParam(value="journalNo", required=false) int journalNo, 
-			ExerciseList exercises) {	    
-	    
-	    service.insertExercises(exercises);
-	   
-	    return "redirect:/journalDetail?journalNo=" + journalNo; 
-	}*/
-	
-	/*
-	@PostMapping("/eWriteProcess") 
-	public String insertExercise(
-			@RequestParam(value="journalNo", required=false) int journalNo, 
-			ArrayList<Exercise> exercises) {	    
-	    
-	    service.insertExercises(exercises);
-	   
-	    return "redirect:/journalDetail?journalNo=" + journalNo; 
-	}
-	*/
-	
-	@PostMapping("/eWriteProcess") 
-	public String insertExercise(String exerciseCate, String exerciseName,
-	        @RequestParam(value="journalNo", required=false) int journalNo, 
-	        @RequestParam(value="exerciseWeight", required=false) int exerciseWeight, 
-	        @RequestParam(value="exerciseSet", required=false) int exerciseSet, 
-	        @RequestParam(value="exerciseRep", required=false) int exerciseRep) { 
-	    Exercise exercise = new Exercise();
-	    exercise.setExerciseCate(exerciseCate);
-	    exercise.setExerciseName(exerciseName);
-	    exercise.setExerciseWeight(exerciseWeight);
-	    exercise.setExerciseSet(exerciseSet); 
-	    exercise.setExerciseRep(exerciseRep);
-	    exercise.setFK_exercise_journal(journalNo);
-	    
-	    service.insertExercise(exercise);
-	   
-	    return "redirect:/journalDetail?journalNo=" + journalNo; 
-	}
-	 
-	
 	@PostMapping("/jWriteProcess")
-	public String insertJournal(String journalTitle,
-			@RequestParam(value = "no", required=false, defaultValue = "1") int no) {
+	public String insertJournal(String journalTitle,			
+			HttpServletRequest request, HttpSession session ,
+			@RequestParam(value = "no", required=false) Integer no) {
+		
+		session = request.getSession();
+		Member member = (Member) session.getAttribute("member");
+		if(member == null){
+			return "redirect:loginForm";
+		}
+		no= member.getUserNo();
+		
 		
 		Journal journal = new Journal();
 		journal.setJournalTitle(journalTitle);
@@ -90,14 +89,35 @@ public class JournalController {
 	
 	@RequestMapping(value = "/trainingJournal", method = RequestMethod.GET)
 	public String getJournal(Model model,
-		@RequestParam(value = "no", required=false, defaultValue = "1") int no) {
+			HttpServletRequest request, HttpSession session ,
+		@RequestParam(value = "no", required=false) Integer no) {
+		
+		session = request.getSession();
+		System.out.println("session:" +session);
+		Member member = (Member) session.getAttribute("member");
+		System.out.println("member:" +member);
+		if(member == null){
+			return "redirect:loginForm";
+		}
+		no= member.getUserNo();
+		System.out.println("no:" +no);
+	    
 		Map<String, Object> modelMap = service.journalList(no);
 		model.addAllAttributes(modelMap);
+		System.out.println("model:" +model);
 		return "trainingJournal";
 	}
 
 	@GetMapping("/journalDetail")
-	public String getExercise(Model model, int journalNo) {
+	public String getExercise(Model model, int journalNo,
+			HttpServletRequest request, HttpSession session) {
+		
+		session = request.getSession();
+		Member member = (Member) session.getAttribute("member");
+		if(member == null){
+			return "redirect:loginForm";
+		}
+		
 		Map<String, Object> modelMap = service.exerciseList(journalNo);
 		model.addAllAttributes(modelMap);
 		model.addAttribute("journalNo", journalNo);
@@ -106,8 +126,14 @@ public class JournalController {
 	
 	
 	@GetMapping("/exerciseWrite")
-	public String getDetail(Model model, int journalNo) {
+	public String getDetail(Model model, int journalNo,
+			HttpServletRequest request, HttpSession session) {
 		
+		session = request.getSession();
+		Member member = (Member) session.getAttribute("member");
+		if(member == null){
+			return "redirect:loginForm";
+		}
 		
 		Journal journal = service.getJounal(journalNo);
 		model.addAttribute(journal);
@@ -125,8 +151,16 @@ public class JournalController {
 	}*/
 
 	@RequestMapping("/delete")
-	public String deleteExercises(@RequestParam("selectedItems") int[] selectedItems, int journalNo) {
-	    for (int exerciseNo : selectedItems) {
+	public String deleteExercises(@RequestParam("selectedItems") int[] selectedItems, int journalNo,
+			HttpServletRequest request, HttpSession session) {
+		
+		session = request.getSession();
+		Member member = (Member) session.getAttribute("member");
+		if(member == null){
+			return "redirect:loginForm";
+		}
+		
+		for (int exerciseNo : selectedItems) {
 	        service.deleteExercise(exerciseNo);
 	    }
 	    return "redirect:journalDetail?journalNo=" + journalNo;
